refactor: restructure firebase configuration and package names
- Migrate package name from example packages to com.the360ghar.stays
- Reorganize Firebase configuration by environment (dev, staging, prod)
- Add Firebase setup documentation and configuration examples
- Remove hardcoded google-services.json files
- Update build.gradle.kts with new package structure
