--- android/app/src/dev/google-services.json
@@ -1,48 +1,0 @@
-{
-  "project_info": {
-    "project_number": "866676409773",
-    "project_id": "stays-app-52ca6",
-    "storage_bucket": "stays-app-52ca6.firebasestorage.app"
-  },
-  "client": [
-    {
-      "client_info": {
-        "mobilesdk_app_id": "1:866676409773:android:e7434edd703cfea8e14d24",
-        "android_client_info": {
-          "package_name": "com.example.stays_app"
-        }
-      },
-      "oauth_client": [],
-      "api_key": [
-        {
-          "current_key": "AIzaSyBcOC2oanpFWvLmIqlsKxNm81LSnyKieeQ"
-        }
-      ],
-      "services": {
-        "appinvite_service": {
-          "other_platform_oauth_client": []
-        }
-      }
-    },
-    {
-      "client_info": {
-        "mobilesdk_app_id": "1:866676409773:android:14b96bc01de5ab69e14d24",
-        "android_client_info": {
-          "package_name": "com.example.stays_app.dev"
-        }
-      },
-      "oauth_client": [],
-      "api_key": [
-        {
-          "current_key": "AIzaSyBcOC2oanpFWvLmIqlsKxNm81LSnyKieeQ"
-        }
-      ],
-      "services": {
-        "appinvite_service": {
-          "other_platform_oauth_client": []
-        }
-      }
-    }
-  ],
-  "configuration_version": "1"
-}
