--- android/app/src/main/kotlin/com/example/stays_app/MainActivity.kt
@@ -1,5 +1,0 @@
-package com.example.stays_app
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity : FlutterActivity()
